rename: AddTemplateContextProvider -> TemplateStore
to better capture its purpose.
