+++ src/features/createAndEditTemplate/workflow/store/TemplateStore.tsx
@@ -30,7 +30,7 @@ export const useTemplateRequestState = () => useContext(TemplateRequestState);
 const TemplateRequestDerivedState = createContext<TemplateRequestDerivedStateType>(initialDerived);
 export const useTemplateRequestDerivedState = () => useContext(TemplateRequestDerivedState);
 
-export const AddTemplateContextProvider = ({ children }: { children: ReactNode }) => {
+export const TemplateStore = ({ children }: { children: ReactNode }) => {
   const [selectedArchitecture, setArchitecture] =
     useState<FirstEmpty<AllowedArchitecture>>(undefined);
   const [selectedOSVersion, setOSVersion] = useState<FirstEmpty<AllowedOSVersion>>(undefined);
